Fix extension loading on initial startup
