+++ packages/shared/src/utils/index.ts
@@ -1,4 +1,6 @@
+export * from './ensure-extensions-dirs';
+export * from './generate-extensions-entry';
+export * from './get-extensions';
 export * from './list-folders';
-export * from './load-extensions';
 export * from './pluralize';
 export * from './resolve-package';
